refactor(passengers): update action imports to new folder structure
